Very big update. AKDump added. String kernel is 100% working. Shell is almost done. All existing functions work but it is to be developed.
I added new .h file: AKDump. It's made for comfortable dumping and a
king of debugging. I think about moving my AKDebug project into this
file. But still it is only on the first part of creating though modified
dump and print work. About AKString: I have tested kernel very well and
I had fixed so many bugs so it defenitely has to work normally (I'm not
pretty sure). As for the shell, it works 'cause in shell there is
nothing that can be broken (Im so naive ye?) so I still develop it but
you can use it without any fear that your program will be broken (no
AHAH0h0hHAhah0Ha)